Add fake sensor manager
This is useful to identify what dependencies the code has on the sensor connector interface.
It also makes it possible to test some parts of the UI without sensors.
However there is still a dependence on the sequence of events from the Sensor Connector,
so any real testing still needs to be done with real sensors.