Allow sensor-tag to take BLE device as argument
Allows caller to pass in a device or for the sensor tag manager to present the Chrome BLE dialoig to user for device selection.  Should improve backwards compatibility so sensor tag devices works with our new general "connect to wireless" button or the previous path where the manager showed a list of devices that only included sensor tags.