Handle GDX sensor disconnection through means other than sensor-interactive UI
A sensor device may become disconnected via means other than the disconnection button in the sensor-interactive UI.  For example, the device might be physically turned off by the user.  If, for any reason, we we detect a device disconnection that is not initiated by the disconnect button, post a warning to the user and complete the disconnection process (remove live reading, remove sensor from pulldown menu, etc.).
[#164439994]