fixes to allow connection to LPSTK SensorTag
The data format has changed however so the humidity and temperature readings from this
code are not correct.
From what I can tell the Humidity service now just returns 32 humidity.
And the IR Temperature service is used to return 32 ambient temperature.
However the calibration for this ambient temperature is very likely different than the IR temperature calibration in this code.