first pass at simplification of components/app.tsx
the components/fixed-display-demo is a directly copy of components/app.tsx
and then several parts were removed from it which are not needed for just
the numeric display.
The intention of this approach is to provide a second example of an app that is
using the sensor manager interface, so we can see about moving some of the
common parts into a common library.

The commit just strips down the extra stuff.