Wire up existing sensor interactive app component [#181112352]
This wires the existing sensor app into the new AP hosted interactive using the fake sensor.

The interactive state data is provided to both the runtime and report and the setter is provided to the runtime.  At this point though that interactive state is ignored pending discussion on how to save data.