Added heartbeat sensor reading [#181112413]
This adds the start heartbeat suport for the sensors so they immediately display pauseable sensor values when connected and not just when collecting data.

There are FIXME comments in the sensor connector, sensor tag and thermoscope managers to add heartbeat support when there are devices available to develop and test with.  For now those handlers show a console error and the device reading falls back to the sensor value.