fix light and color sensor
This sensor was causing the sensor interactive to crash.
It returned multiple sensor types, and most of them had units which are not known. This resulted in an undefined description for the sensor.
The undefined description then caused the code to crash later on.

The solution is identify these unknown unit columns and skip them.

Additionally I cleaned up the handling of the "no interface" case. In this case matchSensorsToDataColumns was called but all it did was reset the two sensors slots.
I checked older versions of the sensor-interactive and found they behaved this same way. In other words this resetting of the sensor slots is not a new behavior.